+++ src/main/scala/org/fs/checker/provider/GenProvider.scala
@@ -7,6 +7,9 @@ trait GenProvider {
   /** Settings prefix for this provider */
   def providerKey: String
 
+  /** Connection timeout */
+  def timeoutMs: Int = 60 * 1000
+
   /** Whether or not the given URL can be checked by this provider */
   def recognizeUrl(url: String): Boolean
 }